bodyprog split, move default controller data to C
* Split bodyprog at func_80033548
* Move default controller data to C
* Move func_80032CE8 to a new file to make its string literal end up in
  the right place.
