+++ include/bodyprog/bodyprog.h
@@ -1727,11 +1727,6 @@ extern s_FsImageDesc g_MainImg0; // 0x80022C74 - TODO: Part of main exe, move to
 
 extern const s_MapType g_MapTypes[16];
 
-extern char D_8002510C[]; // "\aNow_loading."
-
-/** Default key bindings. Multiple configs probably? */
-extern s_ControllerConfig D_8002511C[];
-
 extern u8 const g_12x16FontWidths[];
 
 extern u32 g_MapMsg_Colors[];
