+++ src/bodyprog/bodyprog_800314EC.c
@@ -753,11 +753,3 @@ void GameFs_SaveLoadBinLoad() // 0x80032CA8
     Fs_QueueStartReadTim(FILE_TIM_SAVELOAD_TIM, FS_BUFFER_1, &g_ItemInspectionImg);
     Fs_QueueStartRead(FILE_VIN_SAVELOAD_BIN, FS_BUFFER_1);
 }
-
-void func_80032CE8()
-{
-    // TODO: D_8002510C .rodata comes after the rodata in `globals.c`
-    // This func (and maybe any related non-rodata funcs above) should probably be moved to `bodyprog_80032D1C.c`
-    Gfx_StringSetPosition(SCREEN_POSITION_X(33.75f), SCREEN_POSITION_Y(43.5f));
-    Gfx_StringDraw(D_8002510C, 100); // TODO: .rodata "\aNow_loading."
-}
